Changed query so that it only gets objects that match criteria
Instead of getting all documents and throwing away the ones that don't
match the sessions for the filter, it now only downloads the documents
that match the sessions for the filter. This should help for queries
that have large amounts of data, if you can partition them in some way.